P14 backend

+++ controllers/cards.js
@@ -2,24 +2,28 @@ const { URL } = require('url');
 const mongoose = require('mongoose');
 const { ValidationError } = require('mongoose').Error;
 const Card = require('../models/card');
+const BadRequest = require('../middlewares/handleError');
+const NotFound = require('../middlewares/handleError');
+const Forbidden = require('../middlewares/handleError');
+const InternalServerError = require('../middlewares/handleError');
 
-// Обработчик для получения всех карточек
-const getCards = (req, res) => {
+// Получения всех карточек
+const getCards = (req, res, next) => {
   Card.find()
     .then((cards) => res.json(cards))
-    .catch(() => res.status(500).json({ message: 'Ошибка при получении списка карточек' }));
+    .catch(() => next(new InternalServerError('Ошибка при получении списка карточек')));
 };
 
-// Обработчик для создания карточки
-const createCard = (req, res) => {
+// Создания новой карточки
+const createCard = (req, res, next) => {
   const { name, link } = req.body;
 
   if (!name || !link) {
-    return res.status(400).json({ message: 'Поля name и link обязательны для заполнения' });
+    return next(new BadRequest('Поля name и link обязательны для заполнения'));
   }
 
   if (name.length < 2 || name.length > 30) {
-    return res.status(400).json({ message: 'Длина имени карточки должна быть от 2 до 30 символов' });
+    return next(new BadRequest('Длина имени карточки должна быть от 2 до 30 символов'));
   }
 
   let isValidURL;
@@ -31,7 +35,7 @@ const createCard = (req, res) => {
   }
 
   if (!isValidURL) {
-    return res.status(400).json({ message: 'Поле link должно быть валидным URL' });
+    return next(new BadRequest('Поле link должно быть валидным URL'));
   }
 
   return Card.create({ name, link, owner: req.user._id })
@@ -43,24 +47,24 @@ const createCard = (req, res) => {
     })
     .catch((error) => {
       if (error instanceof ValidationError) {
-        return res.status(400).json({ message: 'Ошибка валидации карточки' });
+        return next(new BadRequest('Ошибка валидации карточки'));
       }
-      return res.status(500).json({ message: 'Ошибка при создании карточки' });
+      return next(new InternalServerError('Ошибка при создании карточки'));
     });
 };
 
-// Обработчик для добавления лайка карточке
-const likeCard = (req, res) => {
+// Добавления лайка карточке
+const likeCard = (req, res, next) => {
   const { cardId } = req.params;
 
   if (!mongoose.Types.ObjectId.isValid(cardId)) {
-    return res.status(400).json({ message: 'Невалидный id карточки' });
+    throw new BadRequest('Невалидный id карточки');
   }
 
   return Card.findByIdAndUpdate(cardId, { $addToSet: { likes: req.user._id } }, { new: true })
     .then((card) => {
       if (!card) {
-        return res.status(404).json({ message: 'Карточка не найдена' });
+        throw new NotFound('Карточка не найдена');
       }
       const {
         _id, name, link, owner, likes,
@@ -69,21 +73,21 @@ const likeCard = (req, res) => {
         _id, name, link, owner, likes,
       });
     })
-    .catch(() => res.status(500).json({ message: 'Ошибка при добавлении лайка карточке' }));
+    .catch(() => next(new InternalServerError('Ошибка при добавлении лайка карточке')));
 };
 
-// Обработчик для удаления лайка с карточки
-const dislikeCard = (req, res) => {
+// Удаления лайка с карточки
+const dislikeCard = (req, res, next) => {
   const { cardId } = req.params;
 
   if (!mongoose.Types.ObjectId.isValid(cardId)) {
-    return res.status(400).json({ message: 'Невалидный id карточки' });
+    throw new BadRequest('Невалидный id карточки');
   }
 
   return Card.findByIdAndUpdate(cardId, { $pull: { likes: req.user._id } }, { new: true })
     .then((card) => {
       if (!card) {
-        return res.status(404).json({ message: 'Карточка не найдена' });
+        throw new NotFound('Карточка не найдена');
       }
       const {
         _id, name, link, owner, likes,
@@ -92,30 +96,34 @@ const dislikeCard = (req, res) => {
         _id, name, link, owner, likes,
       });
     })
-    .catch(() => res.status(500).json({ message: 'Ошибка при удалении лайка с карточки' }));
+    .catch(() => next(new InternalServerError('Ошибка при удалении лайка с карточки')));
 };
 
-// Обработчик для удаления карточки по идентификатору
-const deleteCard = (req, res) => {
+// Функция для удаления карточки
+const deleteCard = (req, res, next) => {
   const { cardId } = req.params;
 
   if (!mongoose.Types.ObjectId.isValid(cardId)) {
-    return res.status(400).json({ message: 'Невалидный id карточки' });
+    throw new BadRequest('Невалидный id карточки');
   }
 
-  return Card.findByIdAndRemove(cardId)
+  return Card.findById(cardId)
     .then((card) => {
       if (!card) {
-        return res.status(404).json({ message: 'Карточка не найдена' });
+        throw new NotFound('Карточка не найдена');
       }
-      const {
-        _id, name, link, owner, likes,
-      } = card;
-      return res.status(200).json({
-        _id, name, link, owner, likes,
-      });
+
+      if (card.owner.toString() !== req.user._id) {
+        throw new Forbidden('У вас нет прав на удаление этой карточки');
+      }
+
+      return Card.findByIdAndRemove(cardId)
+        .then((deletedCard) => {
+          res.status(200).json({ message: 'Карточка успешно удалена', data: deletedCard });
+        })
+        .catch(() => next(new InternalServerError('Ошибка при удалении карточки')));
     })
-    .catch(() => res.status(500).json({ message: 'Ошибка при удалении карточки' }));
+    .catch(() => next(new InternalServerError('Ошибка при поиске карточки')));
 };
 
 module.exports = {
